refactor: sortBy type String -> Enum
