--- src/main/java/com/fourthread/ozang/module/domain/feed/dto/dummy/SortDirection.java
@@ -1,5 +1,0 @@
-package com.fourthread.ozang.module.domain.feed.dto.dummy;
-
-public enum SortDirection {
-  ASCENDING, DESCENDING
-}
