+++ src/main/java/com/fourthread/ozang/module/domain/feed/entity/SortBy.java
@@ -1,0 +1,5 @@
+package com.fourthread.ozang.module.domain.feed.entity;
+
+public enum SortBy {
+  likeCount, createdAt
+}
