+++ src/main/java/com/fourthread/ozang/module/domain/feed/entity/SortDirection.java
@@ -1,0 +1,5 @@
+package com.fourthread.ozang.module.domain.feed.entity;
+
+public enum SortDirection {
+  ASCENDING, DESCENDING
+}
