+++ src/main/java/com/fourthread/ozang/module/domain/user/repository/custom/UserCustomRepository.java
@@ -1,6 +1,6 @@
 package com.fourthread.ozang.module.domain.user.repository.custom;
 
-import com.fourthread.ozang.module.domain.feed.dto.dummy.SortDirection;
+import com.fourthread.ozang.module.domain.feed.entity.SortDirection;
 import com.fourthread.ozang.module.domain.user.dto.response.UserCursorPageResponse;
 import com.fourthread.ozang.module.domain.user.dto.type.Role;
 import java.util.UUID;
